Rename ivtools.utility to ivtools.utils, and silence ivtools.sdm warnings (#1048)
* rename utility to utils, silence some warnings

* rename test_utility.py

* squelch more
